+++ apps/decodex/src/recovery/tests/stale_active/support.rs
@@ -1,0 +1,17 @@
+mod progress_evidence;
+mod release_audit;
+mod runtime_telemetry;
+
+pub(in crate::recovery::tests::stale_active) use self::{
+	progress_evidence::{
+		append_app_server_no_progress_failure_evidence, append_harness_outcome_with_pr_progress,
+		append_harness_outcome_with_review_progress,
+		append_harness_outcome_with_validation_progress, append_no_diff_guardrail_event,
+		append_no_diff_guardrail_event_with_source_error_class, append_phase_goal_recovery_event,
+	},
+	release_audit::{append_stale_active_release_audit, append_stale_active_release_audit_for_run},
+	runtime_telemetry::{
+		append_dead_process_interrupt_control_telemetry, seed_dead_orphan_runtime_telemetry,
+		seed_dead_orphan_runtime_telemetry_without_control_channel,
+	},
+};
